Fix display and deletion(kinda)
Deletion works now and attempts to change the display numbers. However, we still need to address the issue of adding a chop dot between/in front of exisitng chop dots. I want to keep the dots in numerical order from left to right, so in the case we need to place the dot in front of other dots, we need to update the indices of every dot to the right of this new one.